#486: unify the loader
