+++ AtlasUI/AtlasUI/AtlasUIViewController.swift
@@ -9,6 +9,12 @@ public class AtlasUIViewController: UIViewController {
     let mainNavigationController: UINavigationController
     private let atlasReachability = AtlasReachability()
 
+    private let loaderView: LoaderView = {
+        let view = LoaderView()
+        view.hidden = true
+        return view
+    }()
+
     init(atlasCheckout: AtlasCheckout, forProductSKU sku: String) {
         let sizeSelectionViewController = SizeListSelectionViewController(checkout: atlasCheckout, sku: sku)
         mainNavigationController = UINavigationController(rootViewController: sizeSelectionViewController)
@@ -25,6 +31,23 @@ public class AtlasUIViewController: UIViewController {
         view.addSubview(mainNavigationController.view)
         mainNavigationController.view.fillInSuperView()
         atlasReachability.setupReachability()
+    }
+
+}
+
+extension AtlasUIViewController {
+
+    func showLoader() {
+        loaderView.removeFromSuperview()
+        UIApplication.topViewController()?.view.addSubview(loaderView)
+        loaderView.fillInSuperView()
+        loaderView.buildView()
+        loaderView.show()
+    }
+
+    func hideLoader() {
+        loaderView.hide()
+        loaderView.removeFromSuperview()
     }
 
 }
