+++ AtlasUI/AtlasUI/CheckoutSummaryActionsHandler.swift
@@ -28,10 +28,10 @@ extension CheckoutSummaryActionsHandler {
     internal func handleBuyAction() {
         guard let viewController = self.viewController else { return }
 
-        viewController.displayLoader { done in
+        LoaderView.displayLoader { hideLoader in
             viewController.checkout.client.createCheckoutCart(for: viewController.checkoutViewModel.selectedArticleUnit,
             addresses: viewController.checkoutViewModel.selectedAddresses) { result in
-                done()
+                hideLoader()
                 guard let (checkout, cart) = result.process() else { return }
 
                 let checkoutViewModel = CheckoutViewModel(
@@ -49,7 +49,7 @@ extension CheckoutSummaryActionsHandler {
 
     internal func createOrder(forCheckoutId checkoutId: String) {
         guard let viewController = self.viewController else { return }
-        viewController.displayLoader { hideLoader in
+        LoaderView.displayLoader { hideLoader in
             viewController.checkout.client.createOrder(checkoutId) { result in
                 hideLoader()
                 guard let order = result.process() else { return }
@@ -75,9 +75,9 @@ extension CheckoutSummaryActionsHandler {
     private func generateCheckout(customer: Customer) {
         guard let viewController = self.viewController else { return }
 
-        viewController.displayLoader { done in
+        LoaderView.displayLoader { hideLoader in
             viewController.checkout.createCheckoutViewModel(fromModel: viewController.checkoutViewModel) { result in
-                done()
+                hideLoader()
                 guard var checkoutViewModel = result.process() else { return }
 
                 checkoutViewModel.customer = customer
@@ -143,9 +143,9 @@ extension CheckoutSummaryActionsHandler {
         guard let viewController = self.viewController else { return }
         guard Atlas.isUserLoggedIn() else { return loadCustomerData() }
 
-        viewController.displayLoader { done in
+        LoaderView.displayLoader { hideLoader in
             viewController.checkout.client.addresses { result in
-                done()
+                hideLoader()
                 guard let addresses = result.process() else { return }
                 let addressSelectionViewController = AddressPickerViewController(checkout: viewController.checkout,
                     initialAddresses: addresses.map { $0 },
@@ -164,9 +164,9 @@ extension CheckoutSummaryActionsHandler {
         guard let viewController = self.viewController else { return }
         guard Atlas.isUserLoggedIn() else { return loadCustomerData() }
 
-        viewController.displayLoader { done in
+        LoaderView.displayLoader { hideLoader in
             viewController.checkout.client.addresses { result in
-                done()
+                hideLoader()
                 guard let addresses = result.process() else { return }
                 let addressSelectionViewController = AddressPickerViewController(checkout: viewController.checkout,
                     initialAddresses: addresses.filter { $0.pickupPoint == nil } .map { $0 },
