+++ AtlasUI/AtlasUI/LoaderView.swift
@@ -3,6 +3,7 @@
 //
 
 import UIKit
+import AtlasSDK
 
 class LoaderView: UIView {
 
@@ -25,6 +26,14 @@ class LoaderView: UIView {
     internal func hide() {
         activityIndicator.stopAnimating()
         hidden = true
+    }
+
+    internal static func displayLoader(block: (() -> Void) -> Void) {
+        let atlasUIViewController: AtlasUIViewController? = try? Atlas.provide()
+        atlasUIViewController?.showLoader()
+        block {
+            atlasUIViewController?.hideLoader()
+        }
     }
 
 }
